make bin/setup clone to .., not ../commnity
dradis-ce is a 'community' repo itself, so it should live in the same
directory as the other community gems.

bin/setup was actually broken before this as Gemfile.plugins.template is
already trying to load repos from `..`, not `../community`, so running
bin/setup then trying to start the app would give an error as the cloned
gems could not be found.