Issues#index - add toolbar and split in partials
Instead of having the weird toolbar in the center of the screen, we've
moved to a Gmail-style toolbar where different button groups are shown
depending on whether there are any items selected.

The button groups also disappear after a successful Tag/Delete
operation.

We've also split the main view into a number of different partials for
better structuring of the code.