Nodes#show - special treatment to the :services property
We want the 'services' property to always render as a table, so users
can rely on it to create their reports.

Unfortunately when a host has a single port associated with it, the
standard node properties mechanism would store :services as a simple
value instead of an array. This commit patches the rendering code to
make sure :services are always displayed in a table.

This commit closes securityroots/dradispro-tracker#336