NodeProperties - dedup Hash properties
When we set a hash value as a property (by itself or inside an Array),
ActiveModel's serialization (JSONWithIndifferentAccess) will end up
stringifying the keys. Which works as expected.

The problem is when we try to re-set the property to the same value
(which can happen for instance if we upload the same Nmap file twice).
In this case, NodeProperties is pulling a stringified Hash from the
database, but the plugin is sending a hash with symbols as keys, and the
previous logic wouldn't detect the differenc.

This commit calls #with_indifferent_access on the new value we're trying
to store before doing the comparison with the values already stored.

This commit combined with [i] closes dradis/dradis-ce#43

[i]
https://github.com/dradis/dradis-nmap/commit/e2e562197fa8ea83063b934421f052d966556578