Issues#index - add column list to cache key
If a new Issue is added that brings in a extra field not already present
in the project, the original implementation would result in the already
existing issue rows being read from the cache. These rows don't have the
new field and as a result the table gets all out of wack.