fix CatchJsonParseErrors middleware insertion
upgrading to Rails 5 broke this because the ActionDispatch::ParamsParser
middleware doesn't exist anymore. Insert it before a different piece of
middleware instead.

It doesn't really matter where it's inserted as long as it's a) later in
the chain than the code which can raise a ParseError and b) middleware
that's present in all 3 environments.

Rails 5 also deprecates passing the new middleware when as a string; we
now pass in the class constant directly