fix test that fails on CI
on Codeship, `Node.order('created_at ASC').last(3)` sometimes (not
always)returns the *parent* node (the one with label=192.168.0.1), which
makes the test fail. Presumably this is because Codeship is (ironically)
running so fast that the parent node and its children have the same
created_at timestamp, so Postgres can't reliably put them in the right
order.

Change the test so that it doesn't rely on the timestamps to test for
the correct behaviour.