evidence toolbar: fix non-deterministic failing spec
This spec was failing for some seeds, e.g.:

rspec ./spec/features/node_pages/evidence_table_toolbar_spec.rb --seed 6159

The problem is (was) that the first element in the table isn't
necessarily the first item in the `items` array. So an evidence was
being deleted as expected, but the spec then checked if a different
evidence had been deleted.