add data-turbolinks false to comment anchor link
Without this attribute, clicking the link triggers a full turbolinks
page reload rather than just jumping to the correct spot on the page
like you'd expect with a normal HTML same-page anchor tag.

This is a known issue with turbolinks and it doesn't look like
it's going to be fixed soon (the issue linked below is still open
afte two years); in the meantime the solution is to add this data
attribute.

See:

  https://github.com/turbolinks/turbolinks/issues/75