Define service to manage node merges
This service class will aid in the merging of nodes. It's designed to move all
nested children, notes, evidence, issues, and attachments. It will do it's best
to recover from errors including putting any moved files back where they belong.

This service specifically only manages the merging of nodes as data, and not
the merging of nodes as a feature. With this in mind the service does not remove
the original node. It does offer an interface to pass a block which can be used
to remove the node at the time of calling. The block will run within the error
handling for best results.