Add a copy_to method for attachments
Attachment.copy_to will make an identical copy of itself for a target node. The
target node can also be the attachments parent node. In the event of naming
collisions the copy with increment with _copy-## like adding a new attachment.

We've purposfully avoided creating a new instance of Attachment until the new
file exists to avoid creating tmp files. Instead we opt to copy the source file
entirely and then find it as a representation of an attachment.