Simplify the attachment merging
Previously the method was to *Move* existing attachments over to the target
node. If an error occured we *Moved* them back to the source.

Once we account for possible naming collisions we're no longer just moving a
file but we're changing it. If an error occurs we'de have to track the changes
rename the file back, and move it.

Instead we've implemented a new solution. Copy the source files to the target.
Track the files we've moved in *this* merge. If an error occurs simply remove
all the newely copied files. Source files stay exactly where they are until the
source node is deleted.