Move the transformation of extras to the merger
Instead of having the node perform the merge of properties move the logic
back to the Merger. The merger must transform existing extras back into services
to turn back into extras. So we don't create new methods.

This also changes the expected behaviour of the merger. Previously we tried to
keep copies of services that were different. Now we do a Source > Target overwrite
of service related properties.