Scope elements within their own modal
When access elements like #nodes_list unfortunitly there is actually more than one loaded in the dom with the ID. This is because we re-use the same modal, and every modal is pre-rendered in the dom. So we have multiple elements sharing a single id. This meant in the subnode add node form it this function was still finding the add root node elements and not properlly updating the subnode forms. To get around this we scope the elements we reference to the modal the event was triggered from.