Refactor tour methods
We've got a lot of complexity here for what I think is an unused feature.
The current method missing and respond_to methods allow for us to dynamically assign tour names. This isn't actually something we do in the code and I think it's only tested as a result of the implementation. Instead we can build the same methods from the list of frozen tours and remove the dynamic nature of the list.