Inject css into pages that require JS
We inject extra css into pages using JS, or pages that need the animations removed.
Selenium/chroma/capybara often doesn't wait for animations to finish before moving to the next step. This can cause intermittant problems with elements in following steps.

If we remove all animations we solve the issue. The injection needs to happen on every page that's visited so we're overloading the visit function. Just injecting the js as a before action doesn't guarantee it won't be remove/overidden later on.