Add a default user with the welcome setup data
Previously we did not need to create a default user with the welcome date.
With a recent [activities change](https://github.com/dradis/dradis-ce/commit/fce25affc1739071af519efee535e0bef19a7d3f) we moved from using a denormalized email value to a relation. The relation can not be null and requires a valid id. Create a default user along with the welcome setup to satisfy the requirement.

This also solves the current issue where the template in the welcome setup can not find a user and the setup errors out.