Pass the event to manage the revision
Without touching the resources anymore we suffered the problem of no difference between an autosaved record and a new update. The only graceful way to manage this is to simply pass in the event at the time the event occurs. If we're autosaving pass an autosave event. If were updating, pass an update event. Then we can always know what kind of state we're in and what we're moving too.