Communicate uploads between fileupload and editor
This builds a bridge between the EditorToolbar and the existing on page file uploader.
The uploader may use native functionality for drag drop, or copy paste, but requires knowledge of where the event came from to find it's way back to the correct toolbar.
Additionally we need a bridge to allow selecting a new upload via prompt and to pass it to the existing uploader.