Don't sleep predetermined lengths
Instead of always sleeping for a set time we can signal that the
debouncing is finshed. Capybara has it's own polling and timeout to wait
for elements to appear. So if we say find a form thats finished
debouncing we can utilize the build in poller, and find the element
within the shortest possible time.