Cache column title keys
As records go up in quantity the dynamic key generation takes it's toll.

Any record type that allows the use of Dradis Fields is inherently
dynamic. This makes drawing the table complicated because we need to
know all possible values for table headers. Focusing on issues as an
example: Currently we load every issue into memory examine the fields
and build the list of known columns.

I ran some bench marking on this query and the amount of records that
need to be scanned are positively correlated to the amount of time it
takes to process the query.

We do lots of caching on the page already but those very cache keys use
this expensive query as their key. Caching won't help if the key is the
problem.

This commit CACHES the expensive query used to later make the keys. I
heard you like caches so I cached your cache key.