Increase evidence content length size
Some users run into errors where the evidence of the scan files far
exceeds that of our defined content length. Increase the allowed content
length for evidence. There's no reason for us not to if users need the
space. We're also not doing it for _all_ fields because it hasn't proven
necessary in places like issues or notes yet.