KitImportJob#perform - extract the copy/unzip to a helper method
In order for the FileUtils.cp_r to work, we need to make sure the source
folder ends in '/.' so we end up copying the contents and not the folder
itself.

The code for the operation is getting a bit complex so we're better off
extracting it to a helper method.