jquery.textile - render Form view as HTML and attach it
Before this change, the server would use RJS to render the HTML form and
then manipulate the DOM to attach it to the editor.

This worked fine in a world with a single editor, but if there are
multiple editors in one page, the server's RJS doesn't know what
jquery.textile originated the request and ended up attaching the HTML
form to all the editors on the page.

This PR moves that responsibility to the editor itself that loads the
view from the server and then attaches it to itself.