Fix Boards Factory - associate with the right project
Before this PR, a Board created via the factory could end up associated
with a Node in a different project. We used create(:project) inside the
node association which was independent from the board's own :project
association.

By using an after-build callback we can ensure that the node belongs to
the same project already setup via the :project association.

We guard this assignment to allow the caller to specify a different node
(e.g. a host in the project).