Show dropdown when clicking the Nodes '+'' icon
Adds a new behavior to the '+'' next to the Nodes header in the
navigation sidebar. When clicking the '+', the user will see a dropdown
with the link 'Add top-level node'.

When clicking the 'Add top-level node' link, the user will see a modal
with the form to add a new top-level node.

See:
  https://trello.com/c/GTtTdP7C