Prompts#show - render more of the response form the action
Before we were defering a lot of rendering to the Job and Turbo, and
some times we would miss the Turbo updates because Cable hadn't updated
the DOM yet.

With this commit we're rendering all the basic HTML placeholders, and
the Turbo events will always have a place to be rendered on the page.