SessionsController should inherit from SetupRequired
This is to ensure you're not attempting to log in without having run
through the Setup Wizard yet. It's a rare situation that can happen
either in Development or if in the Sandbox environment your DB gets
nuked while you were signed in. Then your session expires, and you're
sent to /login on a pristine DB, which before this PR the controller
wouldn't detect.

After this change, even in that scenario, /login redirects to /setup if
the DB becomes pristine.

The problem is that our Specs were relying on "login trickery" to bypass
the Setup Wizard rather than addressing the problem head on. This PR
also creates the necessary Fixtures to effectively bypass the Wizard for
all tests.