Setup::Kits#create - always mark_as_done
Before this change, if the user chose to upload a Kit, we'd enqueue the
job and assume it'd populate the DB, which would be detected by the
Setup Wizard. Then we'd redirect to /login

However, often, the /login page would detect the pristine DB (because
the background worker hadn't kicked in yet), and send the user back to
Setup::Kit#new.